Fix package imports
- Change eslint plugins for linting imports
- Move core code out of the index file and into its own file. This is so
that the index file can be used strictly for exporting
- Add exports for all measures from the index file in definitions
- Update the README to reflect how imports work now
