+++ .eslintrc.js
@@ -11,7 +11,8 @@ module.exports = {
     'plugin:@typescript-eslint/eslint-recommended',
     'plugin:@typescript-eslint/recommended',
     'prettier',
-    'plugin:ordered-imports/recommended',
+    'plugin:import/recommended',
+    'plugin:import/typescript'
   ],
   parserOptions: {
     ecmaVersion: 12,
@@ -20,27 +21,10 @@ module.exports = {
   plugins: [
     'jest',
     'prettier',
-    'ordered-imports',
+    'import',
   ],
   rules: {
     'prettier/prettier': 2,
-    'ordered-imports/ordered-imports': [
-      'error',
-      {
-        'symbols-first': true,
-        'declaration-ordering': ['type', {
-          ordering: ['side-effect', 'default', 'namespace', 'destructured'],
-          secondaryOrdering: ['source', 'lowercase-last'],
-        }],
-        'specifier-ordering': 'lowercase-last',
-        'group-ordering': [
-          { name: 'project root', match: '^@', order: 20 },
-          { name: 'parent directories', match: '^\\.\\.', order: 30 },
-          { name: 'current directory', match: '^\\.', order: 40 },
-          { name: 'third-party', match: '.*', order: 10 },
-        ],
-      },
-    ],
   },
   parser: '@typescript-eslint/parser',
 };
