+++ .gitignore
@@ -9,3 +9,4 @@ scratch
 
 coverage
 lib
+convert-units-*.tgz
