+++ .prettierignore
@@ -1,2 +1,3 @@
-build
 converage
+lib
+.eslintrc.js
